move assignment out of main into its own file
