+++ main.py
@@ -1,4 +1,6 @@
+
 import store
+from objects import Assignment
 from flask import Flask, render_template, make_response
 
 
@@ -9,19 +11,6 @@ try:
     pstore.load()
 except:
     pstore.make_file()
-
-class Assignment():
-    def __init__(self, name, due, length):
-        self.name = name
-        self.due = due
-        self.length = length
-
-    def __str__(self):
-        return "[Assignment {} {} {}]".format(self.name, self.due, self.length)
-
-    def html(self):
-        return "<p>{}</p>".format(str(self))
-
 
 @app.route("/")
 def home():
